move Message to protobuf package
